# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Health Diet Online Search",
      "description": "A community for Health experts and users"
    },
    "host": "localhost:3000",
    "basePath": "/api/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/experts/register": {
        "post": {
          "tags": ["Experts"],
          "summary": "Create new expert in system",
          "description": "",
          "operationId": "addExpert",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "expert object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Register"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "expert created successfully"
            },
            "400": {
              "description": "expert already exists"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/login":{
        "post": {
          "tags": ["Experts"],
          "summary": "Login expert in system",
          "description": "",
          "operationId": "LoginExpert",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "expert object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Login"
              }
            }
          ],
          "responses": {
            
            "200": {
              "description": "Login successfully"
            },
            "400": {
              "description": "Invalid email or password"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/forgot": {
        "post": {
          "tags": ["Experts"],
          "summary": "sends password reset link to expert email.",
          "description": "",
          "operationId": "ResetPassword",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "expert object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Reset"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "password reset email sent successfully to your email"
            },
            "404": {
              "description": "expert not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/reset": {
        "post": {
          "tags": ["Experts"],
          "summary": "sends password reset link to expert email.",
          "description": "",
          "operationId": "ResetPassword",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "password reset email sent successfully to your email"
            },
            "404": {
              "description": "expert not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/reset/{token}": {
        "post": {
          "tags": ["Experts"],
          "summary": "Resets password for expert",
          "description": "",
          "operationId": "ResetPassword",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "token",
              "description": "token need to in params",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "password",
              "description": "expert password that needs to be updated",
              "required": true,
              "type" : "string"
            }
          ],
          "responses": {
            "200": {
              "description": "password reset successfully"
            },
            "404": {
              "description": "expert not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/get/me": {
        "get": {
          "tags": ["Experts"],
          "summary": "Get expert profile details of login expert",
          "description": "",
          "operationId": "getExpertProfile",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "expert fetched successfully",
              "schema": {
                "$ref": "#/definitions/Expert"
              }
            },
            "404": {
              "description": "expert not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/get/verify":{
        "get": {
          "tags": ["Experts"],
          "summary": "sends email verification email to expert email",
          "description": "",
          "operationId": "VerifyExpert",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "verification email sent successfully to your email"
            },
            "404": {
              "description": "invalid token"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/verify/{token}":{
        "get": {
          "tags": ["Experts"],
          "summary": "Verify expert email",
          "description": "",
          "operationId": "VerifyExpert",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "expert verified successfully"
            },
            "404": {
              "description": "expert not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/experts/update": {
        "put": {
          "tags": ["Experts"],
          "summary": "update profile of the current expert",
          "description": "",
          "operationId": "updateExpertProfile",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "expert object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Expert"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This expert doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
...
        "Expert": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "dob": {
              "type":"date",
              "required":false
            },
            "gender": {
              "type":"string",
              "required":false
            },
            "expertIn": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "required":false
            } ,
            "bio": {
              "type":"string",
              "required":false
            },
            "twitterUrl": {
              "type":"string",
              "required":false
            },
            "instagramUrl": {
              "type":"string",
              "required":false
            },
            "linkedInUrl": {
              "type":"string",
              "required":false
            },
            "location": {
              "type":"string",
              "required":false
            }
            
          }
        },
        "Register":{
          "type": "object",
          "properties": {
            "name": {
              "type":"string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        },
        "Login":{
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        },
        "Reset":{
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            }
          }
        },
...
      } 
}